Remove unwanted default files and move BotService to a Bot layer
