+++ DiscordBot/Bot/BotService.cs
@@ -6,7 +6,7 @@ using Discord.Commands;
 using Discord.WebSocket;
 using Microsoft.Extensions.Configuration;
 
-namespace DiscordBot.Data {
+namespace DiscordBot.Bot {
     public class BotService {
         private readonly IConfiguration _config;
         private DiscordSocketClient _client;
