+++ DiscordBot/Pages/Bot.razor
@@ -1,6 +1,5 @@
 ﻿@page "/bot"
-
-@using DiscordBot.Data
+@using DiscordBot.Bot
 @inject BotService DiscordBot
 
 <h1>Bot Service</h1>
