+++ DiscordBot/Startup.cs
@@ -1,4 +1,3 @@
-using DiscordBot.Data;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Hosting;
@@ -10,6 +9,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using DiscordBot.Bot;
 
 namespace DiscordBot
 {
@@ -26,7 +26,6 @@ namespace DiscordBot
         public void ConfigureServices(IServiceCollection services) {
             services.AddRazorPages();
             services.AddServerSideBlazor();
-            services.AddSingleton<WeatherForecastService>();
             services.AddSingleton<BotService>();
         }
